move layer controls to separate component
